Create web configuration file.
Now this class responsible for web configuration, implements WebMvcConfiguer, create beans for messages, localResolver and localChangeInterceptor
